refactor: promote lib/ids to a folder module
