+++ lib/ids/index.ts
@@ -1,0 +1,1 @@
+export { generateId } from './generate'
